Переименовал классы согласно заданию
